Add configs and scripts for converting v1.x models to v2.0
